requirements/requirements.txt
pandas==1.5.3
numpy==1.22.0
google-oauth2-tool==0.0.3
google-api-python-client==2.79.0
prophet==1.1.2
matplotlib==3.7.0
google-cloud-bigquery==3.6.0
db-dtypes==1.0.5
Jinja2==3.1.2
cssutils==2.6.0
premailer==3.10.0
google-cloud-secret-manager==2.16.0
statsmodels==0.13.5
pmdarima==2.0.3
openpyxl==3.1.2
pyyaml==6.0